% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_annot.R
\name{make_annot}
\alias{make_annot}
\title{A function to make adjcany matrix from edge list and find isolated modules in it}
\usage{
make_annot(adjmat)
}
\arguments{
\item{adjancy_matrix:}{a matrix P X P}
}
\description{
A function to make adjcany matrix from edge list and find isolated modules in it
}
